--- lib/position.rb
@@ -1,7 +1,0 @@
-class Position
-  attr_accessor :x, :y
-  def initialize(x:, y:)
-    self.x = x
-    self.y = y
-  end
-end
