+++ lib/robot.rb
@@ -1,57 +1,59 @@
-require 'geometry'
+##
+# This robot class is aware of where it is, where it is facing, how it moves and how it turns
+#
+# It has "angle" as a base attribute for a more generic approach to movement and can possibly be extended to face in all angles and not just cardinal directions.
 
 class Robot
-  def initialize(step_distance: default_step_distance)
-    self.step_distance = step_distance
-  end
+  RIGHT_ANGLE_IN_DEGREES = 90
+  DIRECTION_TO_TRANSLATION = {
+    NORTH: Translation.new(0, 1),
+    EAST: Translation.new(1, 0),
+    SOUTH: Translation.new(0, -1),
+    WEST: Translation.new(-1, 0)
+  }.freeze
 
-  def place(position, orientation)
-    self.position = position
-    self.direction = Geometry.polar_angle_from_cardinal_direction(orientation)
-  end
+  attr_reader :angle
+  attr_accessor :position
 
-  def placed?
-    [position, orientation].all?
-  end
-
-  def orientation
-    Geometry.cardinal_direction_from_polar_angle(direction)
-  end
-
-  def x_position
+  def x
     position.x
   end
 
-  def y_position
+  def y
     position.y
   end
 
+  def placed?
+    !!position && !!direction
+  end
+
   def turn_left
-    set_direction(direction + turning_angle)
+    self.angle = angle - RIGHT_ANGLE_IN_DEGREES
   end
 
   def turn_right
-    set_direction(direction - turning_angle)
+    self.angle = angle + RIGHT_ANGLE_IN_DEGREES
+  end
+
+  def direction=(direction_name)
+    self.angle = CardinalDirection.from_name(direction_name)
+  end
+
+  def direction
+    CardinalDirection.name(angle)
   end
 
   def position_ahead
-    polar_translation_vector = Geometry::PolarVector.new(polar_angle: direction, magnitude: step_distance)
-    polar_translation_vector.translate_position(position)
+    position.translate(translation)
   end
 
   private
 
-  attr_accessor :step_distance, :position, :direction
-
-  def default_step_distance
-    1
+  def translation
+    DIRECTION_TO_TRANSLATION[direction]
   end
 
-  def turning_angle
-    90 # degrees
-  end
-
-  def set_direction(direction)
-    self.direction = direction % 360
+  def angle=(angle)
+    @angle = angle % 360
   end
 end
