--- lib/robot_commands.rb
@@ -1,11 +1,0 @@
-require 'robot_commands/turn_left'
-require 'robot_commands/turn_right'
-require 'robot_commands/report'
-require 'robot_commands/move'
-require 'robot_commands/place'
-
-module RobotCommands
-  def self.all
-    [TurnLeft, TurnRight, Report, Move, Place]
-  end
-end
