--- lib/robot_commands/turn_left.rb
@@ -1,13 +1,0 @@
-module RobotCommands
-  class TurnLeft
-    def self.can_handle?(user_command)
-      user_command == "LEFT"
-    end
-
-    def apply(_user_command, robot)
-      return unless robot.placed?
-
-      robot.turn_left
-    end
-  end
-end
