+++ lib/translation.rb
@@ -1,0 +1,4 @@
+##
+# This Translation class holds the delta values to offset a coordinate on a two-dimensional plane
+
+Translation = Struct.new(:x, :y)
